knowledge_portal_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
elasticsearch==8.11.0
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic>=2.0.1
email-validator==2.1.0
pydantic-settings==2.0.3
python-dotenv==1.0.0
motor==3.3.1
aiohttp==3.9.1
python-slugify==8.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0
psycopg2-binary==2.9.9
uuid==1.30
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
pre-commit==3.5.0
google-auth>=2.40.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
aioredis>=2.0.0
requests>=2.31.0 